+++ lib/gamefic/scene/custom.rb
@@ -1,7 +1,0 @@
-module Gamefic
-  # A Custom Scene allows for complete configuration of its behavior upon
-  # instantiation. It is suitable for direct instantiation or subclassing.
-  #
-  class Scene::Custom < Scene::Base
-  end
-end