+++ lib/gamefic/world/playbook.rb
@@ -118,35 +118,16 @@ module Gamefic
         syn
       end
 
-      # Get an array of actions, derived from the specified command, that the
-      # actor can potentially execute.
-      # The command can either be a single string (e.g., "examine book") or a
-      # list of tokens (e.g., :examine, @book).
+      # Get a Dispatcher to select actions that can potentially be executed
+      # from the specified command string.
       #
-      # @return [Array<Gamefic::Action>]
-      def dispatch(actor, *command)
-        result = []
-        result.concat dispatch_from_params(actor, command[0], command[1..-1]) if command.length > 1
-        result.concat dispatch_from_string(actor, command.join(' ')) if result.empty?
-        result
-      end
-
-      # Get an array of actions, derived from the specified command, that the
-      # actor can potentially execute.
-      # The command should be a plain-text string, e.g., "examine the book."
-      #
-      # @return [Array<Gamefic::Action>]
-      def dispatch_from_string actor, text
-        result = []
+      # @param actor [Actor]
+      # @param text [String]
+      # @return [Dispatcher]
+      def dispatch(actor, text)
         commands = Syntax.tokenize(text, actor.syntaxes)
-        commands.each do |c|
-          actions_for(c.verb).each do |a|
-            next if a.hidden?
-            o = a.attempt(actor, c.arguments)
-            result.unshift o unless o.nil?
-          end
-        end
-        sort_and_reduce_actions result
+        actions = commands.flat_map { |cmd| actions_for(cmd.verb).reject(&:hidden?) }
+        Dispatcher.new(actor, commands, sort_and_reduce_actions(actions))
       end
 
       # Get an array of actions, derived from the specified verb and params,
@@ -156,10 +137,7 @@ module Gamefic
       def dispatch_from_params actor, verb, params
         result = []
         available = actions_for(verb)
-        available.each do |a|
-          result.unshift a.new(actor, params) if a.valid?(actor, params)
-        end
-        sort_and_reduce_actions result
+        Dispatcher.new(actor, [Command.new(verb, params)], sort_and_reduce_actions(available))
       end
 
       # Duplicate the playbook.
@@ -218,7 +196,7 @@ module Gamefic
       end
 
       def sort_and_reduce_actions arr
-        arr.sort_by.with_index { |a, i| [a.rank, -i]}.reverse.uniq(&:class)
+        arr.sort_by.with_index { |a, i| [a.rank, i] }.reverse.uniq
       end
     end
   end