+++ spec/plot_spec.rb
@@ -45,6 +45,15 @@ describe Gamefic::Plot do
     expect(plot.in_subplot?(actor)).to be(true)
   end
 
+  it 'deletes concluded subplots on update' do
+    plot = Gamefic::Plot.new
+    subplot = plot.branch Gamefic::Subplot
+    expect(plot.subplots).to include(subplot)
+    subplot.conclude
+    plot.update
+    expect(plot.subplots).to be_empty
+  end
+
   it "pauses on a pause scene" do
     plot = Gamefic::Plot.new
     actor = plot.cast Gamefic::Actor
@@ -75,5 +84,14 @@ describe Gamefic::Plot do
     actor.cue plot.default_conclusion
     plot.update
     expect(ran).to be true
+  end
+
+  it 'supports multiple players' do
+    plot = Gamefic::Plot.new
+    player1 = plot.make_player_character
+    plot.introduce player1
+    player2 = plot.make_player_character
+    plot.introduce player2
+    expect(plot.players).to eq([player1, player2])
   end
 end