+++ CHANGELOG.md
@@ -1,2 +1,13 @@
+# 2.1.1 - July 23, 2021
+- Remove gamefic/scene/custom autoload
+
+# 2.1.0 - June 21, 2021
+- Remove redundant MultipleChoice prompt
+- Deprecate Scene::Custom
+
+# 2.0.3 - December 14, 2020
+- Remove unused Index class
+- Active#conclude accepts data argument
+
 # 2.0.2 - April 25, 2020
 - Improved snapshot serialization